third_party/blink/web_tests/external/wpt/workers/constructors/SharedWorker/interface-objects-expected.txt

This is a testharness.js-based test.
FAIL Test if interface objects exist in a shared worker assert_equals: expected "" but got "These were missing: Worker, SharedWorker"
Harness: the test ran to completion.
